Normal execution enabled
The simulator is able to execute workflows such that there can be too many tasks arriving at the same moment to be able to execute them immediately on the nodes.
The io_aware_scheduler_cut sched is yet to be coded.
The characteristics of benchmarked workflows are yet to be inserted.